Ice cold cash

+++ Application/Entities/CartItem.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public int ProductId { get; set; }
         public string ProductName { get; set; }
-        public decimal Price { get; set; }
-        public decimal Count { get; set; }
+        public long Price { get; set; }
+        public int Count { get; set; }
         public int CartId { get; set; }
         public virtual Cart Cart { get; set; }
         public virtual Product Product { get; set; }
